TSMOK-75: Dodano relacje sala przedmiot

+++ classrooms/static/js/classrooms.js
@@ -82,18 +82,22 @@ $(document).ready(function() {
     $("form.add_classroom").submit(function(event) {
         $.post(addURL, $(this).serialize(), function(data){
             var newClassroom = jQuery.parseJSON(data)[0];
-            var newLi = $('<li id ="'+newClassroom.pk+'"><span id ="resource-name-'+newClassroom.pk+'" class="list-element">' +newClassroom.fields.name +'</span><span class="edit" onclick="edit_item('+newClassroom.pk+');">&#x1f589;</span><span id ="resource-delete-'+newClassroom.pk+'" class="close" onclick="delete_item('+newClassroom.pk+', \''+newClassroom.fields.name+'\');">&#x232b;</span></li>').hide();
-            
-            //TODO: sort() nie dziala, nowe elementy sa dodawane na koncu ul#list
-            $("#list").add(newLi.fadeIn(800)).sort(asc_sort).appendTo('#list');
-            //$("#list").add(newLi.fadeIn(800)).sort(sortAlpha).appendTo('#list');
-            
-            $('li#'+newClassroom.pk).on('click', 'span.list-element', function() {
-                $( this ).parent().toggleClass( "checked" );
-            });
-            if ($('#list li').length != 0) {
-                $('p#noResources').remove();
-            }
+            if (newClassroom.error === "UNIQUE_NAME_VIOLATED") {
+                $('#successModal').css('display', 'block').delay(3000).fadeOut('slow');
+            } else {
+                var newLi = $('<li id ="'+newClassroom.pk+'"><span class="row-content"><span id ="resource-name-'+newClassroom.pk+'" class="list-element">' +newClassroom.fields.name +'</span></span><span class="row-content"> <span id="resource-building-'+newClassroom.pk+'" class="list-element">' +newClassroom.fields.building +'</span> </span><span class="row-content"> <span id="resource-seats-'+newClassroom.pk+'" class="list-element">' +newClassroom.fields.seats +'</span> </span> <span class="row-content"> <span id="resource-relsubjects-'+newClassroom.pk+'" class="list-element">'+newClassroom.relatedSubjects+'</span></span> <span class="edit" onclick="edit_item('+newClassroom.pk+');">&#x1f589;</span><span id ="resource-delete-'+newClassroom.pk+'" class="close" onclick="delete_item('+newClassroom.pk+', \''+newClassroom.fields.name+'\');">&#x232b;</span></li>').hide();
+
+                //TODO: sort() nie dziala, nowe elementy sa dodawane na koncu ul#list
+                $("#list").add(newLi.fadeIn(800)).sort(asc_sort).appendTo('#list');
+                //$("#list").add(newLi.fadeIn(800)).sort(sortAlpha).appendTo('#list');
+
+                $('li#'+newClassroom.pk).on('click', 'span.list-element', function() {
+                    $( this ).parent().toggleClass( "checked" );
+                });
+                if ($('#list li').length != 0) {
+                    $('p#noResources').remove();
+                }
+                }
         });
         event.preventDefault();
         $('#addModal').css('display', 'none');
@@ -104,8 +108,15 @@ $(document).ready(function() {
     $("form.edit_classroom").submit(function(event) {
         $.post(editURL, $(this).serialize(), function(data){
             var newClassroom = jQuery.parseJSON(data)[0];
-            $('span#resource-name-'+newClassroom.pk).text(newClassroom.fields.name);
-            $('span#resource-delete-'+newClassroom.pk).attr("onclick",'delete_item('+newClassroom.pk+', \''+newClassroom.fields.name+'\');');
+            if (newClassroom.error === "UNIQUE_NAME_VIOLATED") {
+                $('#successModal').css('display', 'block').delay(3000).fadeOut('slow');
+            } else {
+                $('span#resource-name-'+newClassroom.pk).text(newClassroom.fields.name);
+                $('span#resource-building-'+newClassroom.pk).text(newClassroom.fields.building);
+                $('span#resource-seats-'+newClassroom.pk).text(newClassroom.fields.seats);
+                $('span#resource-relsubjects-'+newClassroom.pk).text(newClassroom.relatedSubjects);
+                $('span#resource-delete-'+newClassroom.pk).attr("onclick",'delete_item('+newClassroom.pk+', \''+newClassroom.fields.name+'\');');
+            }
         });
         event.preventDefault();
         $('#editModal').css('display', 'none');
